Rename googleservice to conservation
googleservice was an inaccurate name for the services provided
