+++ api/views.py
@@ -11,7 +11,7 @@ from FullSend.authhelper import get_signin_url, get_token_from_code, get_access_
 from FullSend.authhelper import api_key_required, post_required, get_token_from_refresh_token
 from mail.outlookservice import get_me, get_my_messages, generate_email, send_message, make_api_call
 from mail.formhandler import MailForm, QuickForm
-from api.googleservice import get_email_template
+from api.conservation import get_email_subject, get_email_template, has_valid_template_type
 from api.models import ServerAuthentication
 
 # Create your views here.
@@ -78,13 +78,17 @@ def formsubmit(request):
 
     if request.body is None:
         return HttpResponseBadRequest('Body is empty')
+
     settings = json.loads(request.body)
+    if not has_valid_template_type(settings['type']):
+        return HttpResponseBadRequest('Type was invalid')
+
     email = {
         'message': {
-            'subject': 'Test email',
+            'subject': get_email_subject('Title'),
             'body': {
                 'contentType': 'HTML',
-                'content': get_email_template('Lol what\'s up my dude this is from Django')
+                'content': get_email_template(settings['type'])
             },
             'toRecipients': [
                 {
